Make membership summary generation reusable for element-ios and others
This separates the code for generating the textual summary of membership change
events into separate files. Having this separate from the react code allows other
platforms, for instance the Element iOS app, to reuse the code without inheriting
the whole of matrix-react-sdk.

Closes: vector-im/element-web#16793
